Rename files
Jira ticket: CAMS-402
